Bug 323732 - An unnecessary error in the error log appears when a closed remote C/C++ project is deleted

+++ org.eclipse.ptp.rdt.core/src/org/eclipse/ptp/internal/rdt/core/index/IndexBuildSequenceController.java
@@ -449,14 +449,9 @@ public class IndexBuildSequenceController implements IResourceChangeListener {
 		if (event.getSource() instanceof IWorkspace) {
 			
 			IResource resource = event.getResource();
-			if (event.getType() == IResourceChangeEvent.PRE_DELETE) {
-				if (resource.getType() == IResource.PROJECT
-						&& RemoteNature.hasRemoteNature((IProject) resource)) {
+			if (event.getType() == IResourceChangeEvent.PRE_DELETE && resource.getType() == IResource.PROJECT) {
 					deleteProjectStaus(resource.getName());
-				}
 			}
-			
-
 		}
 	}
